requirements.txt
httpx>=0.28
pillow>=12.0
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
rembg>=2.0
fastapi>=0.115
uvicorn>=0.34
